[프론트엔드] 220601 장바구니, 주문, 관리자 이슈 해결

See merge request sw_track/class_02_seoul/web_project/team20/pet_shop!75

+++ src/views/adminOrderList/adminOrderList.js
@@ -92,6 +92,7 @@ async function getItemInfo(orderList) {
 
 async function handleCancelBtn() {
   console.log(`취소 버튼 ${this.name}`);
+  if (!confirm('주문 내역을 삭제하시겠습니까?')) return;
   try {
     const res = await Api.delete('/api/order/admin/delete', '', {
       orderId: this.name,
